Created a version of CreativeTabs which caches its contents.
The vanilla implementation of CreativeTabs will loop through every item
in the game when it is initialized, to get a list of all valid items.
This is hardly an issue when you have only a handful of creative tabs
and items/blocks, the more items/blocks you have, and the more tabs you
have, the more cpu cycles are needlessly wasted. This tab caches the
basic Item instances, which does use more memory, however it improves
the over all loading performance, which is a well worth trade off,
seeing as the most complex of modded environments will ever break the
4096MB, but suffer from high CPU ussage.