Allow blocks to determine their pathfinding type
This commit introduces an IBlockHooks interface that serves a similar role
to IForgeBlock or IBlockExtensions. It currently offers a hook for
determining the pathfinding type of a block instead of relying on the
convoluted logic of getBlockPathTypeRaw. While similar hooks already exist
on various modloader platforms this hook is platform independent.

In the future the IBlockHooks interface will be expanded to include more
hooks. The existing hooks such as those in ILightningConductive may also
be migrated to this interface in a future update.